restructuring test cases, removed locale specific
